fix linting in notFound.jsx
